more reorganizing and cleanup
