+++ frontend/src/main.tsx
@@ -1,7 +1,7 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import App from './pages/App'
-import './styles/global.css'
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import App from './components/App';
+import './styles/global.css';
 
 // routing
 import { BrowserRouter } from 'react-router-dom';
